Arreglo de compatibilidad
